fix(resources): pack_format 6 -> 34（1.21.1），合并残留的 data/opencomputers 到 opencomputers_neo；确认 OpenOS loot 完整（assets/opencomputers_neo/loot/openos 179 文件）

+++ src/main/scala/li/cil/oc/client/renderer/block/SmartBlockModelBase.scala
@@ -143,7 +143,8 @@ trait SmartBlockModelBase extends BakedModel {
       for (i <- 0 until rot) {
         val tmp = u; u = v; v = (-(tmp - 0.5)) + 0.5
       }
-      rawData(vertex.x, vertex.y, vertex.z, facing, texture, texture.getU(u * 16), texture.getV(v * 16), colorRGB)
+      // 1.21.1: TextureAtlasSprite.getU/getV 只接受 float。
+      rawData(vertex.x, vertex.y, vertex.z, facing, texture, texture.getU((u * 16).toFloat), texture.getV((v * 16).toFloat), colorRGB)
     })
   }
